separates askQuestion in its own file
